3 - Debugging/agent.py
from typing import Annotated
from typing_extensions import TypedDict
from langgraph.graph import START , END
from langgraph.graph.state import StateGraph
from langgraph.graph.message import add_messages
from langgraph.prebuilt import ToolNode
from langchain_core.tools import tool
from langchain_core.messages import BaseMessage
from langgraph.prebuilt import tools_condition
import os 
from dotenv import load_dotenv

load_dotenv()

os.environ["GROQ_API_KEY"] = os.getenv("GROQ_API_KEY")
os.environ["LANGCHAIN_API_KEY"] = os.getenv("LANGCHAIN_API_KEY")
os.environ["LANGSMITH_TRACING"] = "true"
os.environ["LANGSMITH_PROJECT"] = "Test Project"

from langchain.chat_models import init_chat_model
llm = init_chat_model("groq:qwen/qwen3-32b")

class State(TypedDict):
    messages: Annotated[list[BaseMessage],add_messages]

def make_tool_graph():
    from langchain_core.tools import tool
    from langgraph.prebuilt import ToolNode

    @tool
    def add(a: float, b: float):
        """Add two numbers together."""
        return a + b

    tools = [add]

    tool_node = ToolNode([add])

    llm_with_tool = llm.bind_tools([add])

    def call_llm_model(state: State):
        return {
            "messages": [llm_with_tool.invoke(state["messages"])]
        }
    # Graph
    builder = StateGraph(State)
    builder.add_node("tool_calling_llm", call_llm_model)
    builder.add_node("tools", ToolNode(tools))

    # Add Edges
    builder.add_edge(START,"tool_calling_llm")
    builder.add_conditional_edges(
        "tool_calling_llm", 
        # If the latest message from the llm contains a tool call, go to the tools node, otherwise end the graph execution. 
        tools_condition)

    builder.add_edge("tools", END)

    # Compile The Graph
    graph = builder.compile()
    return graph

tool_agent = make_tool_graph()