pyproject.toml
[project]
name = "mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.8",
    "langgraph-cli[inem,inmem]>=0.4.14",
    "langsmith>=0.7.1",
    "python-dotenv>=1.2.1",
]
